FLUID MECHANICS ARE INNNNN
now i have a fluid map that stores the fluids added by the player ..... then the fluid simulator will iterate thru all the values of the fluidmap to simulate.

on chunk unload the simulated fluid is saved to modifed map and deleted from fluidmap
on chunk load it doesnt go back to fluid map since the data is already simulated.
in future to update the simulation based on block break/placement ill be reloading the fluid data from modified chunk and be running  a live simulation when u are in the chunk.